Rework WebHooks again.
It's important to be able to determine the URL a webhook will
appear at from inside handlers. This isn't easy, as even from within
a mux, you don't know the final URL. But if we want handlers to be
able to talk to their webooks,we need to be able to dine where they
live.

This adds a URL method for webhooks, and a SetURL method. The latter
is called by the mux to tell the web hook where it lives, the former
can be called by fellow handlers to find the external URL.